+++ uikit/src/main/java/com/sendbird/uikit/interfaces/FormSubmitButtonClickListener.kt
@@ -7,7 +7,9 @@ import com.sendbird.android.message.MessageForm
  * Interface definition for a callback to be invoked when the submit button of the form is clicked.
  *
  * @since 3.12.1
+ * @deprecated As of 3.26.0, this feature is no longer supported.
  */
+@Deprecated("As of 3.26.0, this feature is no longer supported.")
 fun interface FormSubmitButtonClickListener {
     /**
      * Called when the submit button of the form is clicked.
@@ -15,6 +17,7 @@ fun interface FormSubmitButtonClickListener {
      * @param message the message that contains the form
      * @param form the form to be submitted
      * @since 3.12.1
+     * @deprecated As of 3.26.0, this feature is no longer supported.
      */
     fun onClicked(message: BaseMessage, form: MessageForm)
 }