+++ uikit/src/main/java/com/sendbird/uikit/internal/interfaces/OnFeedbackRatingClickListener.kt
@@ -7,7 +7,9 @@ import com.sendbird.android.message.FeedbackRating
  * Interface definition for a callback to be invoked when a feedback rating is clicked.
  *
  * @since 3.13.0
+ * @deprecated As of 3.26.0, this feature is no longer supported.
  */
+@Deprecated("As of 3.26.0, this feature is no longer supported.")
 fun interface OnFeedbackRatingClickListener {
     /**
      * Called when a feedback rating is clicked.
@@ -15,6 +17,7 @@ fun interface OnFeedbackRatingClickListener {
      * @param message the message that contains Feedback.
      * @param rating the feedback rating.
      * @since 3.13.0
+     * @deprecated As of 3.26.0, this feature is no longer supported.
      */
     fun onFeedbackClicked(message: BaseMessage, rating: FeedbackRating)
 }